what: jwt authentication using context and improve error handling

+++ resolvers/recipe.resolver.js
@@ -1,15 +1,42 @@
 import RecipeModel from '../models/recipe.model.js';
+import { GraphQLError } from 'graphql';
+import RecipeHelper from '../helpers/recipe.helper.js';
 
 const recipeResolver = {
   Query: {
-    recipe: async (parent, { id }) => {
+    recipe: async (parent, { id }, contextValue) => {
+      console.log('Context is get Single Recipe by Id');
+      console.log(contextValue);
+      if (!contextValue.user) {
+        throw new GraphQLError('User is not authenticated', {
+          extensions: {
+            code: 'UNAUTHENTICATED',
+            http: { status: 401 },
+          },
+        });
+      }
       const recipe = await RecipeModel.findById(id);
-      return recipe;
+      if (!recipe) {
+        return {
+          __typename: 'NotExistsError',
+          message: `Recipe with id ${id} does not exists.`,
+        };
+      }
+      return {
+        __typename: 'Recipe',
+        ...recipe._doc,
+        id: recipe._id,
+      };
     },
-    async getRecipes(parent, args, context) {
-      console.log(context);
-      console.log(context.authorization);
-      console.log('called get recipes');
+    async getRecipes(parent, args, contextValue) {
+      if (!contextValue.user) {
+        throw new GraphQLError('User is not authenticated', {
+          extensions: {
+            code: 'UNAUTHENTICATED',
+            http: { status: 401 },
+          },
+        });
+      }
       const amount = args.amount;
       const allRecipes = await RecipeModel.find()
         .sort({ createdAt: -1 })
@@ -21,7 +48,19 @@ const recipeResolver = {
   Mutation: {
     // destructure the args object to get recipeInput object and then destructure
     // recipeInput to get the name and description properties
-    createRecipe: async (parent, { recipeInput: { name, description } }) => {
+    createRecipe: async (
+      parent,
+      { recipeInput: { name, description } },
+      contextValue
+    ) => {
+      if (!contextValue.user) {
+        throw new GraphQLError('User is not authenticated', {
+          extensions: {
+            code: 'UNAUTHENTICATED',
+            http: { status: 401 },
+          },
+        });
+      }
       const createdRecipe = new RecipeModel({
         name: name,
         description: description,
@@ -36,22 +75,79 @@ const recipeResolver = {
       };
     },
 
-    deleteRecipe: async (_, { id }) => {
+    deleteRecipe: async (_, { id }, contextValue) => {
+      if (!contextValue.user) {
+        throw new GraphQLError('User is not authenticated', {
+          extensions: {
+            code: 'UNAUTHENTICATED',
+            http: { status: 401 },
+          },
+        });
+      }
+      const isExists = await RecipeHelper.isRecipeExists(id);
+      if (!isExists) {
+        return {
+          __typename: 'NotExistsError',
+          message: `Recipe with id ${id} does not exists.`,
+        };
+      }
       const isDeleted = (await RecipeModel.deleteOne({ _id: id })).deletedCount;
-      return isDeleted; // return true if something is deleted, 0 if nothing is deleted
+      return {
+        __typename: 'RecipeSuccess',
+        isSuccess: isDeleted, // return true if something is deleted, 0 if nothing is deleted
+        message: 'Recipe deleted.',
+      };
     },
 
-    editRecipe: async (_, { id, recipeInput: { name, description } }) => {
+    editRecipe: async (
+      _,
+      { id, recipeInput: { name, description } },
+      { user }
+    ) => {
+      if (!user) {
+        throw new GraphQLError('User is not authenticated', {
+          extensions: {
+            code: 'UNAUTHENTICATED',
+            http: { status: 401 },
+          },
+        });
+      }
+      const isExists = await RecipeHelper.isRecipeExists(id);
+      if (!isExists) {
+        return {
+          __typename: 'NotExistsError',
+          message: `Recipe with id ${id} does not exists.`,
+        };
+      }
       const isEdited = (
         await RecipeModel.updateOne(
           { _id: id },
           { name: name, description: description }
         )
       ).modifiedCount;
-      return isEdited; // return 1 if something is edited, 0 if nothing is edited
+      return {
+        __typename: 'RecipeSuccess',
+        isSuccess: isEdited, // return 1 if something is edited, 0 if nothing is edited
+        message: 'Recipe Edited.',
+      };
     },
 
-    incrementThumbsUp: async (_, { id }) => {
+    incrementThumbsUp: async (_, { id }, { user }) => {
+      if (!user) {
+        throw new GraphQLError('User is not authenticated', {
+          extensions: {
+            code: 'UNAUTHENTICATED',
+            http: { status: 401 },
+          },
+        });
+      }
+      const isExists = await RecipeHelper.isRecipeExists(id);
+      if (!isExists) {
+        return {
+          __typename: 'NotExistsError',
+          message: `Recipe with id ${id} does not exists.`,
+        };
+      }
       await RecipeModel.findByIdAndUpdate(
         { _id: id },
         {
@@ -59,10 +155,30 @@ const recipeResolver = {
         },
         { new: true }
       );
-      return true;
+
+      return {
+        __typename: 'RecipeSuccess',
+        isSuccess: true,
+        message: 'Thumbs up incremented..',
+      };
     },
 
-    incrementThumbsDown: async (_, { id }) => {
+    incrementThumbsDown: async (_, { id }, { user }) => {
+      if (!user) {
+        throw new GraphQLError('User is not authenticated', {
+          extensions: {
+            code: 'UNAUTHENTICATED',
+            http: { status: 401 },
+          },
+        });
+      }
+      const isExists = await RecipeHelper.isRecipeExists(id);
+      if (!isExists) {
+        return {
+          __typename: 'NotExistsError',
+          message: `Recipe with id ${id} does not exists.`,
+        };
+      }
       await RecipeModel.findByIdAndUpdate(
         { _id: id },
         {
@@ -70,7 +186,11 @@ const recipeResolver = {
         },
         { new: true }
       );
-      return true;
+      return {
+        __typename: 'RecipeSuccess',
+        isSuccess: true,
+        message: 'Thumbs Down incremented..',
+      };
     },
   },
 };
